refac(halo2): move msm unittests to `msm.rs`
